remotion_sub/pyproject.toml
[project]
name = "remotion-sub"
version = "0.1.0"
description = "Article to explainer video tool using Remotion + LLM + TTS"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "claude-code-sdk>=0.0.1",
    "click>=8.0.0",
    "aiofiles>=24.0.0",
    "aiosqlite>=0.20.0",
    "python-multipart>=0.0.20",
    "websockets>=14.0",
    "sse-starlette>=2.0.0",
]

[project.scripts]
remotion-sub = "remotion_sub.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/remotion_sub"]
